ARM: mstar: Add mstar prefix to all of the dtsi/dts files
Based on Arnd's comment[0] all of the MStar dtsi/dts files should
have had a prefix.

This moves the files, fixes the Makefile that generates dtbs and
fixes up the MAINTAINERS entry.

Fixing up some includes in the files themselves is left for a later
commit as rolling it into this commit resulted in a confusing diff.

0 - https://lore.kernel.org/linux-arm-kernel/CAK8P3a0maQhfaerwG4KgFZOrUPwueKOp2+MOeG9C=+8ZNzc2Kg@mail.gmail.com/

Link: https://lore.kernel.org/r/20201002133418.2250277-5-daniel@0x0f.com
Signed-off-by: Daniel Palmer <daniel@0x0f.com>
Signed-off-by: Olof Johansson <olof@lixom.net>

[ upstream commit: f1ef5b78a76224ee741925abad0a4b3af482d57f ]