dt-bindings:iio:impedance-analyzer:adi,ad5933 yaml conversion.
The example in this one had a completely wrong compatible so I've
fixed that. Otherwise, a fairly simple conversion.

Note the driver itself is still in staging.  Looking back at the
last discussion around this, I think we were just waiting for some
test results on some refactors.  As such the binding should be stable
even if the driver might need a little more love and attention.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Cc: Marcelo Schmitt <marcelo.schmitt1@gmail.com>
Cc: Gabriel Capella <gabriel@capella.pro>
Cc: Alexandru Ardelean <Alexandru.Ardelean@analog.com>
Link: https://lore.kernel.org/r/20201031184854.745828-6-jic23@kernel.org

[ upstream commit: f97cee494dc92395a668445bcd24d34c89f4ff8c ]