dt-bindings:iio:light:upisemi,us51882: txt to yaml conversion.
I don't have an up to date address for Adriana Reus so I've put myself
as the binding maintainer for this one.  I'm happy to hand over to Adriana
or anyone else who wants take it on!

This has a lot of optional tuning parameters. The docs are modified to try
and put the default values in the description of each one rather than a
forwards reference to the example.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20201031184854.745828-20-jic23@kernel.org

[ upstream commit: 6032afbdfb6164cfacd0721381b0802fabf64df9 ]