dt-bindings:iio:magnetometer:bosch,bmc150_magn: txt to yaml conversion.
This describes the bindings for both stand along magnetometers and ones
which form part of a multi chip package.
Given original author hasn't been active remotely recently I've
put myself as maintainer for this one.  I would of course like to
hand this over to someone more appropriate so shout out if this is you!

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20201031184854.745828-26-jic23@kernel.org

[ upstream commit: ab6eeb9c76b2e1584714d5a5ebfcc6d9d5a4c47c ]