ARM: dts: s5pv210: Drop unneeded io-channel-ranges property in Aries
This property is only useful if a node has children as it allows them
to then use io-channel properties in the parent.  Here there are no
children.

This is harmless, but we are planning to shortly drop this property
as it is rarely used correctly and there is little reason it would
ever be needed as we can just provide the io-channels property to
any child nodes that need it.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Link: https://lore.kernel.org/r/20201115192951.1073632-8-jic23@kernel.org
Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>

[ upstream commit: 0f8159371c225a7b019b612a5c101e8b839c6c46 ]