arm64: dts: rockchip: Light "sys" LED on NanoPi R2S
Set NanoPi R2S's "sys" LED to be on by default. This matches the
behaviour of the stock FriendlyWRT image, and makes it much easier
to tell when the thing has finished booting.

Suitable triggers for the two network LEDs cannot realistically be
configured from DT, so leave them be.

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Link: https://lore.kernel.org/r/f066be60aa99460a45d04113c5e507d6602186f1.1611187213.git.robin.murphy@arm.com
Signed-off-by: Heiko Stuebner <heiko@sntech.de>

[ upstream commit: 833821eeab916631d0cd22dbcda8db4b2a3acc3d ]