ARM: dts: mmp3-dell-ariel: Add the power button node
This adds support for the power button attached to the Embedded Controller
on a Dell Wyse 3020 "Ariel" board.

However, while the EC itself is controlled via I2C, the input capability
for the power button acts as a separate device attached to the SPI, hence
it has a separate device node.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>
Link: https://lore.kernel.org/r/20210121034130.1381872-11-lkundrak@v3.sk'
Signed-off-by: Arnd Bergmann <arnd@arndb.de>

[ upstream commit: 49061130eb1f654065b25903d57dad41519ee6c9 ]