ARM: dts: blanche: Configure pull-up for SOFT_SW and SW25 GPIO keys
The GPIO pins connected to the 4 Software Switches (SW2) and the second
Tact Switch (SW25) do not have external pull-up resistors, but rely on
internal pull-ups being enabled.  Fortunately this is satisfied by the
initial state of these pins.

Make this explicit by enabling bias-pull-up, to remove the dependency on
initial state and/or boot loader configuration.

Note that the GPIO pin connected to the first Tact Switch (SW24) does
have an external pull-up resistor.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Link: https://lore.kernel.org/r/011e4c461767f2dd690b655b3dd501eb554184c1.1619785905.git.geert+renesas@glider.be

[ upstream commit: 1f27fedead91eb6077c299a98ea3d9fe2f9955db ]