dt-bindings: iio: dac: adi,ad5764: Add missing binding document
This binding covers the ad5744, ad5744r, ad5764 and ad5764r DACs.
Note that the driver currently assumes the internal reference is used for
the r parts. The binding as defined relaxes this constraint.

There is no support in the binding or driver for the two digital IO
signals.  I do not propose to add that until we have a means to
test any such binding.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Cc: Lars-Peter Clausen <lars@metafoo.de>
Reviewed-by: Rob Herring <robh@kernel.org>
Acked-by: Nuno Sá <nuno.sa@analog.com>
Link: https://lore.kernel.org/r/20210627163244.1090296-13-jic23@kernel.org

[ upstream commit: 96e137558d5823787a410c3c89d076fdc8b8c720 ]