dt-bindings: phy: brcm,ns-usb2-phy: bind just a PHY block
The old binding was covering the whole DMU block space (DMU block
contains CRU block which contains USB PHY). It was a bad design,
overkill and a non-generic solution.

Northstar's USB 2.0 PHY is a small block (part of the CRU MFD) and
binding should be designed to represent that properly. Rework the
binding to map just PHY with the "reg" property and use syscon to
reference shared register that controls block access.

The old binding is deprecated now.

Signed-off-by: Rafał Miłecki <rafal@milecki.pl>
Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20210913080024.6951-2-zajec5@gmail.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>

[ upstream commit: 55b9b741712de84d144cfed6cb3987314002ea21 ]