dt-bindings: w1-gpio: Drop redundant 'maxItems'
Drop another redundant maxItems that snuck into w1-gpio binding.

If a property has an 'items' list, then a 'minItems' or 'maxItems' with the
same size as the list is redundant and can be dropped. Note that is DT
schema specific behavior and not standard json-schema behavior. The tooling
will fixup the final schema adding any unspecified minItems/maxItems.

Fixes: dd2c898bc20b ("dt-bindings: w1: Convert 1-Wire GPIO binding to a schema")
Cc: Daniel Mack <zonque@gmail.com>
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: 71a9aa162d7b2f9b92b51462203324bcd5a5aeff ]