ARM: dts: stm32: Add support for the emtrion emSBC-Argon
This patch presents the DT patches for the emtrion GmbH
Argon board series. They are available with STM32MP157
from STMicroelectronics with 512 MByte Memory.

The devicetree stm32mp157c-emstamp-argon.dtsi is the common part
providing the module components and the basic support for the SoC.
The support for the emSBC-Argon baseboard in the developer-kit
configuration is provided by the stm32mp157c-emsbc-argon.dts file.

Signed-off-by: Reinhold Mueller <reinhold.mueller@emtrion.com>
Signed-off-by: Alexandre Torgue <alexandre.torgue@foss.st.com>

[ upstream commit: 16e3e44c5b874e07dabcf2e9fd5527d810cbecdc ]