dt-bindings: remoteproc: ti: Add mailbox provider nodes to example
In order to make the 'mboxes' property in the TI remoteproc examples
parseable, mailbox provider nodes are needed. Normally, the examples
have a __fixup__ node which can be used for determining each
phandle+arg entry. However, for this binding the arg cells contain a
phandle, and the __fixups__ information can't be used.

Signed-off-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20220119181053.3846613-1-robh@kernel.org

[ upstream commit: 24a4b57a3d014751ec8ba8b33f38390d5bda94c2 ]