dt-bindings: interconnect: Convert snoc-mm to a sub-node of snoc
snoc and snoc-mm in downstream use the same address space a mistake which
we have carried over into upstream. In silicon terms snoc-mm has been
tacked on to the original snoc which came from msm8916.

Convert to a better description where snoc-mm is a sub-node of snoc, thus
avoiding a mmio remap collision when using the RPM interconnect driver.

Cc: Rob Herring <robh+dt@kernel.org>
Cc: devicetree@vger.kernel.org
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20220129032735.2410936-2-bryan.odonoghue@linaro.org
Signed-off-by: Georgi Djakov <djakov@kernel.org>

[ upstream commit: dfeef93fe3eee6b7b5cbe209be3dbad59426f189 ]